[Interop] Fix zerocopy fallback for unsupported dtypes, structs, and test
- Add dtype parameter to can_zerocopy(); only i32/i64/f32/f64/u1 are
  supported by the C++ DLPack implementation -- f16, u8, u16 etc. now
  correctly fall back to the kernel copy path
- Pass dtype through all can_zerocopy() call sites (field.py, matrix.py)
- Add can_zerocopy() gate to Ndarray.to_torch() with kernel copy fallback
  for unsupported dtypes/backends (was previously unconditional DLPack)
- Force copy=True for StructField children since struct SNodes use
  interleaved memory that DLPack cannot represent correctly
- Relax test_fused_kernels assertion: to_torch() may compile 0 kernels
  (zero-copy) or 1 kernel (copy path)

+++ python/quadrants/lang/field.py
@@ -267,7 +267,7 @@ class ScalarField(Field):
             dlpack_to_torch,
         )
 
-        if copy is not True and can_zerocopy(is_field=True, is_scalar_field=True, shape=self.shape):
+        if copy is not True and can_zerocopy(is_field=True, dtype=self.dtype, is_scalar_field=True, shape=self.shape):
             tc = dlpack_to_torch(self)
             if tc.device.type == "cpu":
                 np_arr = tc.numpy()
@@ -306,7 +306,7 @@ class ScalarField(Field):
             dlpack_to_torch,
         )
 
-        if copy is not True and can_zerocopy(is_field=True, is_scalar_field=True, shape=self.shape):
+        if copy is not True and can_zerocopy(is_field=True, dtype=self.dtype, is_scalar_field=True, shape=self.shape):
             import torch  # pylint: disable=C0415
 
             tc = dlpack_to_torch(self)
